requirements.txt
numpy
opencv-python
torch
torchvision
faiss-cpu
tqdm
fastapi
uvicorn
python-multipart
jinja2
aiofiles
yt-dlp
